Move AccountConfirmationMessages and EmailVerified to pages
